Move base-client/index.ts to base-client/base-client.ts
